同步数据为空时bug修复

+++ app/models/activity.rb
@@ -2,11 +2,13 @@ class Activity < ActiveRecord::Base
   attr_accessible :username, :activityname, :name, :phone
 
   def self.update_people_list(username,list)
-    Activity.delete_all(:username => username)
-    people_list = list
-    people_list.each do |l|
-      new_list = Activity.new(l)
-      new_list.save
+    if list
+      Activity.delete_all(:username => username)
+      people_list = list
+      people_list.each do |l|
+        new_list = Activity.new(l)
+        new_list.save
+      end
     end
   end
 end
