+++ app/models/event.rb
@@ -2,11 +2,13 @@ class Event < ActiveRecord::Base
   attr_accessible :username, :activityname
 
   def self.update_events(username,activities)
-    Event.delete_all(:username => username)
-    activity = activities
-    activity.each do |a|
-      new_activity = Event.new(a)
-      new_activity.save
+    if activities
+      Event.delete_all(:username => username)
+      activity = activities
+      activity.each do |a|
+        new_activity = Event.new(a)
+        new_activity.save
+      end
     end
   end
 end
